+++ packages/core-js-compat/targets-parser.js
@@ -18,6 +18,7 @@ const validTargets = new Set([
   'edge',
   'electron',
   'firefox',
+  'hermes',
   'ie',
   'ios',
   'node',