+++ tests/compat/hermes-adapter.mjs
@@ -1,0 +1,3 @@
+const [HERMES_PATH] = argv._;
+
+await $`${ HERMES_PATH } -w -commonjs ./tests/compat`;