+++ tests/compat/rhino-runner.js
@@ -2,5 +2,5 @@ require('./tests');
 require('./compat-data');
 require('./common-runner');
 
-/* eslint-disable no-restricted-globals -- output */
+/* eslint-disable-next-line no-restricted-globals -- output */
 global.showResults('rhino', print);